+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.586.1](https://github.com/lightdash/lightdash/compare/0.586.0...0.586.1) (2023-05-31)
+
+
+### Bug Fixes
+
+* pivot table crashing in minimal page ([#5696](https://github.com/lightdash/lightdash/issues/5696)) ([be1210e](https://github.com/lightdash/lightdash/commit/be1210ec377aefa1800411687aeee8e2191bb790))
+
+# [0.586.0](https://github.com/lightdash/lightdash/compare/0.585.0...0.586.0) (2023-05-31)
+
+
+### Features
+
+* update validation model and table ([#5682](https://github.com/lightdash/lightdash/issues/5682)) ([8a1bc84](https://github.com/lightdash/lightdash/commit/8a1bc840605f03d8ed4d6fd359503cf234b60f7b))
+
 # [0.585.0](https://github.com/lightdash/lightdash/compare/0.584.1...0.585.0) (2023-05-31)